+++ environment.properties
@@ -1,8 +1,8 @@
 # ENVIRONMENT TEST PROPERTIES
 #Dynamic
-username=
-password=
-token=
+username=learner.mlt@gmail.com
+password=hellopivotal123
+token=7fb7281222ce7f5f0d4ef956a0221c62
 #CHROME or FIREFOX
 browser=CHROME
 #Fixed
