--- src/main/java/org/fundacionjala/pivotal/browser/Browser.java
@@ -1,16 +1,0 @@
-package org.fundacionjala.pivotal.browser;
-
-import org.openqa.selenium.WebDriver;
-
-/**
- * Created by pivotal-test Team.
- */
-public interface Browser {
-
-    /**
-     * This method return a Web Driver instance of a specified browser.
-     *
-     * @return the Web Driver instance.
-     */
-    WebDriver getBrowser();
-}
