--- src/main/java/org/fundacionjala/pivotal/browser/DriverFactory.java
@@ -1,32 +1,0 @@
-package org.fundacionjala.pivotal.browser;
-
-import org.openqa.selenium.WebDriver;
-
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Created by pivotal-test Team.
- */
-public final class DriverFactory {
-
-
-    /**
-     * Private Constructor for the DriverFactory utility class.
-     */
-    private DriverFactory() {
-    }
-
-    /**
-     * Return the WebDriver instance specified by the environment properties.
-     *
-     * @param driverType Enum value specified from DriverType.
-     * @return a WebDriver instance.
-     */
-    public static WebDriver getDriverManager(DriverType driverType) {
-        Map<DriverType, Browser> map = new HashMap<>();
-        map.put(DriverType.CHROME, new ChromeBrowser());
-        map.put(DriverType.FIREFOX, new FirefoxBrowser());
-        return map.get(driverType).getBrowser();
-    }
-}
