--- src/main/java/org/fundacionjala/pivotal/browser/DriverType.java
@@ -1,9 +1,0 @@
-package org.fundacionjala.pivotal.browser;
-
-/**
- * Created by pivotal-test Team.
- */
-public enum DriverType {
-    CHROME,
-    FIREFOX
-}
