+++ src/main/java/org/fundacionjala/pivotal/core/browser/DriverType.java
@@ -1,0 +1,9 @@
+package org.fundacionjala.pivotal.core.browser;
+
+/**
+ * Created by pivotal-test Team.
+ */
+public enum DriverType {
+    CHROME,
+    FIREFOX
+}
