--- src/main/java/org/fundacionjala/pivotal/utilities/CommonActions.java
@@ -1,102 +1,0 @@
-package org.fundacionjala.pivotal.utilities;
-
-import org.fundacionjala.pivotal.browser.DriverManager;
-import org.openqa.selenium.JavascriptExecutor;
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-
-import java.util.List;
-
-/**
- * Created by pivotal-test Team.
- */
-public final class CommonActions {
-
-    /**
-     * This is the constructor for the utility class.
-     */
-    private CommonActions() {
-
-    }
-
-    /**
-     * This method set text content to a web element.
-     *
-     * @param webElement Is web element.
-     * @param content    Is the content that will be set to the web element.
-     */
-    public static void setInputField(WebElement webElement, String content) {
-        DriverManager.getInstance().getWebDriverWait().until(ExpectedConditions.visibilityOf(webElement));
-        webElement.clear();
-        webElement.sendKeys(content);
-    }
-
-    /**
-     * This method perform a click action in a web element.
-     *
-     * @param webElement Is the web element that will be pressed.
-     */
-    public static void clickButton(WebElement webElement) {
-        DriverManager.getInstance().getWebDriverWait().until(ExpectedConditions.elementToBeClickable(webElement));
-        webElement.click();
-    }
-
-    /**
-     * This method perform a click in a non visible element in the UI.
-     *
-     * @param webElement the WebElement non visible in the UI.
-     */
-    public static void jsClickElement(WebElement webElement) {
-        ((JavascriptExecutor) DriverManager.getInstance().getWebDriver())
-                .executeScript("arguments[0].click();", webElement);
-    }
-
-    /**
-     * This method verifies if a web element is visible.
-     *
-     * @param webElement is the web element.
-     * @return true if web element is visible or false if it isn't visible.
-     */
-    public static boolean isVisible(WebElement webElement) {
-        try {
-            return webElement.isDisplayed();
-        } catch (NoSuchElementException e) {
-            return false;
-        }
-    }
-
-    /**
-     * This method perform a search in a WebElement list based on a content string parameter.
-     *
-     * @param elements is the WebElements lists.
-     * @param content  is the content parameter.
-     * @return the WebElement search result.
-     */
-    public static WebElement findWebElement(List<WebElement> elements, String content) {
-        return elements.stream()
-                .filter(element -> content.equals(element.getText()))
-                .findAny()
-                .orElse(null);
-    }
-
-    /**
-     * This method return the text content of a WebElement.
-     *
-     * @param webElement is the WebElement to extract the text.
-     * @return the text content of the WebElement.
-     */
-    public static String getTextContent(WebElement webElement) {
-        DriverManager.getInstance().getWebDriverWait().until(ExpectedConditions.visibilityOf(webElement));
-        return webElement.getText();
-    }
-
-    /**
-     * This method get title of current page.
-     *
-     * @return title of the current page.
-     */
-    public static String getPageTitle() {
-        return DriverManager.getInstance().getWebDriver().getTitle();
-    }
-}
