+++ src/test/java/org/fundacionjala/pivotal/stepdef/restapi/RequestSteps.java
@@ -1,15 +1,16 @@
 package org.fundacionjala.pivotal.stepdef.restapi;
 
+import java.util.Map;
+
 import cucumber.api.java.en.And;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.When;
 import io.restassured.response.Response;
-import org.fundacionjala.pivotal.restapi.RequestManager;
-import org.fundacionjala.pivotal.utilities.DataBuilder;
-import org.fundacionjala.pivotal.utilities.Helper;
-import org.fundacionjala.pivotal.utilities.SharedVariableList;
 
-import java.util.Map;
+import org.fundacionjala.pivotal.core.restapi.RequestManager;
+import org.fundacionjala.pivotal.util.DataInterpreter;
+import org.fundacionjala.pivotal.util.Helper;
+import org.fundacionjala.pivotal.util.SharedVariableList;
 
 /**
  * Created by pivotal-test Team.
@@ -32,11 +33,11 @@ public class RequestSteps {
      * Step definition to perform a POST  request.
      *
      * @param param is the specified end point.
-     * @param map   is the body content.
+     * @param map   is the map body content.
      */
     @Given("^^a POST request to \"([^\"]*)\" with the following data$")
     public void aPOSTRequestToWithTheFollowingData(String param, Map<String, String> map) {
-        response = RequestManager.post(DataBuilder.builtEndPoint(param), map);
+        response = RequestManager.post(DataInterpreter.builtEndPoint(param), map);
         helper.setRequestStatus(response.getStatusCode());
         helper.setProjectVariable(map.get("name"));
     }
@@ -48,9 +49,29 @@ public class RequestSteps {
      */
     @When("^a DELETE request to \"([^\"]*)\"$")
     public void aDELETERequestTo(String param) {
-        response = RequestManager.delete(DataBuilder.builtEndPoint(param));
+        response = RequestManager.delete(DataInterpreter.builtEndPoint(param));
         helper.setRequestStatus(response.getStatusCode());
-        System.out.println(helper.getRequestStatus());
+    }
+
+    /**
+     * Step definition to perform a PUT request.
+     *
+     * @param param is the specified end point.
+     * @param map   is the map body content.
+     */
+    @When("^a PUT request to \"([^\"]*)\" with:$")
+    public void aPUTRequestToWith(String param, Map<String, String> map) {
+        response = RequestManager.put(DataInterpreter.builtEndPoint(param), map);
+    }
+
+    /**
+     * Step definition to perform a GET request.
+     *
+     * @param param is the specified end point.
+     */
+    @When("^a GET request to \"([^\"]*)\"$")
+    public void aGETRequestTo(String param) {
+        response = RequestManager.get(DataInterpreter.builtEndPoint(param));
     }
 
     /**
